Distribute support & Backwards compatibility for Python 2.5
The 'json' module is not available in older versions of python. This changeset
adds 'simplejson' as dependency for older versions, and uses it in older
installations.

In order to support package dependencies, I am using "distribute". According to
"The Hitchhiker's Guide to Packaging", this seems to be the best solution until
distutils2 arrives.

I also took the chance to add a stub README file.